DS-12900 Move password management model classes out to a separate project
