+++ password-management/src/main/assemblies/sonatype-bundle.xml
