+++ password-management/src/main/assemblies/zip.xml
